editing main.js and delete database.js
